requirements.txt
sqlalchemy>=2.0
alembic
psycopg2-binary
boto3
psycopg[binary]